refactor: manage logs data to logs page
