+++ src/Components/utils/translations.js
@@ -1915,4 +1915,8 @@ export const translations = {
     RO: "Eroare la încărcarea grupurilor existente",
     RU: "Ошибка при загрузке существующих групп",
   },
+  logs: {
+    RO: "Loguri",
+    RU: "Логи",
+  },
 };
